% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutilate.R
\name{addEdges}
\alias{addEdges}
\alias{removeEdges}
\title{Add or remove edges}
\usage{
addEdges(graph, edges)

removeEdges(graph, edges)
}
\arguments{
\item{graph}{a \code{mixedgraph} object}

\item{edges}{list of edges to be added/removed}
}
\description{
Add or remove edges
}
\details{
At the moment no effort is made to 
detect duplication in addEdges().  To be added later.
Currently removeEdges() forces all edges to be
represented by adjacency matrices.
}
\section{Functions}{
\itemize{
\item \code{removeEdges}: remove edges
}}

