% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutilate.R
\name{addNodes}
\alias{addNodes}
\title{Add additional nodes to a graph}
\usage{
addNodes(graph, k, vnames)
}
\arguments{
\item{graph}{a \code{mixedgraph} object}

\item{k}{the number of nodes to be added}

\item{vnames}{an optional character vector of names}
}
\description{
Add additional nodes to a graph
}
