% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Adjacencies.R
\name{adj}
\alias{adj}
\title{Find adjacent vertices}
\usage{
adj(graph, v, etype, dir = 0, inclusive = TRUE, sort = 1)
}
\arguments{
\item{graph}{an object of class \code{mixedgraph}}

\item{v}{vertices to find adjacencies}

\item{etype}{edge types to consider; defaults to all}

\item{inclusive}{logical indicating whether elements of \code{v} can be
included in output group.}

\item{sort}{integer:1 for unique but unsorted, 2 for
sorted (0 for possibly repeated and unsorted).  If edges are stored as a matrix
then output will always be unique and sorted.}

\item{directed}{for directed edges, indicates which direction to search in:
1: along direction, -1: against direction, 0: both directions.}
}
\description{
Generic function for finding adjacent vertices based on any kind of edge.
}
\details{
The argument \code{directed} is recycled for multiple edge types, but
has no effect for edges without a specified direction.
}
\seealso{
\code{\link{grp}} for paths
}

