% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Edges.R
\name{adjList}
\alias{adjList}
\alias{revAdjList}
\alias{symAdjList}
\title{Get adjacency list of edges}
\usage{
adjList(edges, n, directed = FALSE, transpose = FALSE)

revAdjList(edges)

symAdjList(edges, unq = TRUE)
}
\arguments{
\item{edges}{list or matrix giving edges in the form of an
\code{eList}, \code{adjMatrix} or \code{edgeMatrix}}

\item{n}{total number of vertices (defaults to maximum value)}

\item{directed}{logical: if \code{TRUE} edges are assumed directed}

\item{transpose}{logical: if \code{TRUE} we consider children instead of parents}

\item{unq}{logical: should values be made unique and sorted?}
}
\description{
Get adjacency list of edges
}
\details{
Stores adjacencies by one of their vertices.  E.g. directed edges
are stored indexed by the child vertex, undirected by both neighbours.
}
\section{Functions}{
\itemize{
\item \code{revAdjList()}: Get reverse directions for edges

\item \code{symAdjList()}: Make adjList symmetric

}}
