% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Edges.R
\name{adjMatrix}
\alias{adjMatrix}
\title{Get adjacency matrix from list of edges}
\usage{
adjMatrix(edges, n, directed = FALSE, sparse = FALSE)
}
\arguments{
\item{edges}{list or matrix giving edges}

\item{n}{total number of vertices, defaults to maximal value}

\item{directed}{logical: if \code{TRUE} edges are assumed directed}

\item{sparse}{logical: use a sparse matrix?}
}
\description{
Get adjacency matrix from list of edges
}
