% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Edges2.R
\name{as.adjMatrix}
\alias{as.adjMatrix}
\title{Put edges in correct format}
\usage{
as.adjMatrix(x, valid = FALSE)
}
\arguments{
\item{x}{object containing edge information}

\item{valid}{logical indicating whether format is already correct}
}
\description{
Put edges in correct format
}
\details{
If \code{valid = TRUE} then entry should be a valid adjacency matrix, and checks
are not performed.
}
