% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Edges2.R
\name{as_adjacency}
\alias{as_adjacency}
\alias{as.adjMatrix}
\alias{as.adjList}
\title{Put edges in correct format}
\usage{
as.adjMatrix(x, valid = FALSE, n)

as.adjList(x, valid = FALSE, n)
}
\arguments{
\item{x}{object containing edge information}

\item{valid}{logical indicating whether format is already correct}

\item{n}{number of vertices}
}
\description{
Functions to place adjacency object in a specific format.
}
\details{
If \code{valid = TRUE} then for \code{as.adjMatrix} the argument \code{x} should be a
valid adjacency matrix, and no checks are performed.  Similarly for \code{as.adjList}
the object should be a valid adjacency list.
}
\section{Functions}{
\itemize{
\item \code{as.adjMatrix()}: Convert to an \code{adjMatrix} object

\item \code{as.adjList()}: Convert to an \code{adjList} object

}}
