% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adjacencies.R
\name{barren}
\alias{barren}
\title{Find barren, sterile, orphaned vertices}
\usage{
barren(graph, v = graph$v)
}
\arguments{
\item{graph}{an object of class \code{mixedgraph}}

\item{v}{set of vertices of \code{graph}}
}
\description{
Find barren, sterile, orphaned vertices
}
\details{
Barren vertices (within \code{v}) are those
that have no proper descendants also within \code{v}.
Sterile (orphaned) vertices in \code{v} have no 
children (parents) also within \code{v}.

Need to think about whether thse
do the right thing for \code{v} a proper subset of the 
vertices
}

