% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Family.R
\name{barrenSets}
\alias{barrenSets}
\title{Get barren subsets}
\usage{
barrenSets(
  graph,
  topOrder,
  max_size,
  same_dist = FALSE,
  sort = 1,
  return_anc_sets = FALSE
)
}
\arguments{
\item{graph}{object of class \code{mixedgraph}}

\item{topOrder}{optionally, a topological order}

\item{max_size}{integer giving maximum size to consider}

\item{same_dist}{logical: should barren sets be in the same district?}

\item{sort}{integer:1 for unique but unsorted, 2 for
sorted.}

\item{return_anc_sets}{logical: return ancestral sets for each vertex as an attribute?}
}
\description{
Return list of barren subsets up to specified size
}
\details{
Uses \code{clique} algorithm on a suitable undirected graph.

\strong{Warning:} Doesn't work for cyclic graphs.
}
