% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_algs.R
\name{BK}
\alias{BK}
\title{Bron-Kerbosch Algorithm}
\usage{
BK(R, P, X, nbs, max_len = length(P))
}
\arguments{
\item{R, P, X}{current largest complete set, sets of vertices still to consider, and already considered}

\item{nbs}{list of neighbouring vertices}

\item{max_len}{maximum length of output}
}
\description{
Bron-Kerbosch Algorithm
}
