% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRAIL.R
\name{BRAIL}
\alias{BRAIL}
\title{BRAIL is Block-Randomized Adaptive Iterative Lasso}
\usage{
BRAIL(X, y, family = "gaussian", tau = 0.8, B = 200,
  lasso.control = list(), ridge.control = list())
}
\arguments{
\item{X}{is a list containing k matrices, each of the matrices has n rows. They are the 'blocks'.}

\item{y}{is the response vector of length n.}

\item{family}{is a description of the error distribution and link function to be used in the model. In our package, "binomial", "gaussian" and  "poisson" are available.}

\item{tau}{is the stability cut-off, which is used for choosing the support features in each block.}

\item{B}{is the number of bootstraps in each iteration.}

\item{lasso.control}{is a list of glmLasso related arguments, like KLB, thresh and max.iter.}

\item{ridge.control}{is a list of ridgeLasso related arguments, like lambda, max.iter and thresh.}
}
\value{
a list containing n vectors. Each element of the list will be a vector of length p_k, where p_k is the number of columns in the k-th block.
}
\description{
BRAIL is Block-Randomized Adaptive Iterative Lasso
}
\examples{
X1 <- matrix(rnorm(500), ncol = 10)
X2 <- matrix(rnorm(500), ncol = 10)
y <- rbinom(50, 1, 0.6)
X <- list(X1, X2)
BRAIL(X, y, family = "binomial", tau = 0.8, B = 200, lasso.control= list(KLB = 10, max.iter = 1e6, thresh = 0.005), ridge.control = list(lambda = 1, max.iter = 1e4, thresh = 1e-5))
}
