% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Edges.R
\name{check_edge_type}
\alias{check_edge_type}
\alias{is.edgeMatrix}
\alias{is.eList}
\alias{is.adjMatrix}
\alias{is.adjList}
\title{Check if edge object could be in given format}
\usage{
is.edgeMatrix(object, n, checknm = TRUE)

is.eList(object, n, checknm = TRUE)

is.adjMatrix(object, n, checknm = TRUE)

is.adjList(object, n, checknm = TRUE)
}
\arguments{
\item{object}{item of purported type}

\item{n}{(optionally) number of vertices in graph}

\item{checknm}{logical: use class of object to determine answer?}
}
\description{
Check if edge object could be in given format
}
\section{Functions}{
\itemize{
\item \code{is.edgeMatrix()}: check if could be \code{edgeMatrix}

\item \code{is.eList()}: check if could be \code{eList}

\item \code{is.adjMatrix()}: check if could be \code{adjMatrix}

\item \code{is.adjList()}: check if could be \code{adjList}

}}
