% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutilate.R
\name{chgEnds}
\alias{chgEnds}
\title{Modify edge marks}
\usage{
chgEnds(graph, v1, v2, to = "arrow")
}
\arguments{
\item{graph}{simple graph of class \code{mixedgraph}}

\item{v1, v2}{end points to respectively preserve and change}

\item{to}{one of 'arrow', 'tail', 'circle'}
}
\description{
Modify edge marks
}
