% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Family.R
\name{claudius}
\alias{claudius}
\title{Find Claudius of a (bidirected-connected) set}
\usage{
claudius(graph, v)
}
\arguments{
\item{graph}{a \code{mixedgraph} object}

\item{v}{set of vertices to consider}
}
\description{
Find the Claudius of a (presumably) bidirected-connected set.
}
\details{
Drops strict spouses of \code{v} and any of their 
descendants.
}
\examples{
data(gr1)
claudius(gr1, 1)
claudius(gr1, 4)

}
