% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Construct.R
\name{construct_graphs}
\alias{construct_graphs}
\alias{makeGraphComplete}
\alias{complete_graph}
\alias{makeGraphEmpty}
\alias{makeGraphChain}
\alias{chain_graph}
\alias{makeGraphCycle}
\alias{makeGraphStar}
\alias{makeGraphGrid}
\alias{makeGraphBipartite}
\title{Graph Constructor Functions}
\usage{
makeGraphComplete(n, type = "undirected")

complete_graph(n, type = "undirected", use_cpp = TRUE)

makeGraphEmpty(n)

makeGraphChain(n, type = "undirected")

chain_graph(n, type = "undirected", use_cpp = TRUE)

makeGraphCycle(n, type = "undirected")

makeGraphStar(n, type = "undirected", out = FALSE)

makeGraphGrid(n, m = n, type = "undirected")

makeGraphBipartite(n, m, type = "undirected")
}
\arguments{
\item{n}{number of vertices (or rows for a grid)}

\item{type}{type of edges to use}

\item{use_cpp}{should C++ function be used?  Defaults to \code{TRUE}}

\item{out}{should edges be directed out from the centre?}

\item{m}{number of columns or size of second set}
}
\description{
Graph Constructor Functions
}
\details{
Provides some construction
functions for common graphs: complete graphs,
chains, cycles and stars.
}
\section{Functions}{
\itemize{
\item \code{makeGraphComplete()}: graph with all adjacencies

\item \code{complete_graph()}: complete graph (faster implementation)

\item \code{makeGraphEmpty()}: graph with no edges

\item \code{makeGraphChain()}: graph with chain of edges

\item \code{chain_graph()}: complete graph (faster implementation)

\item \code{makeGraphCycle()}: graph with cycle of edges

\item \code{makeGraphStar()}: star shaped graph

\item \code{makeGraphGrid()}: grid of vertices

\item \code{makeGraphBipartite()}: bipartite graph

}}
