% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Convert.R
\name{convert}
\alias{convert}
\title{Convert graph to format associated with specific package}
\usage{
convert(graph, format = "mixedgraph", cur_format)
}
\arguments{
\item{graph}{graphical object, of one of classes listed below}

\item{format}{character string giving new format to convert to}
}
\description{
Convert graph to format associated with specific package
}
\details{
Currently limited functionality, only
converts 'ADMGs' to 'mixedgraph' objects.
}

