% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convert.R
\name{determinePackage}
\alias{determinePackage}
\title{Determine which graph package evaluation will
occur in.}
\usage{
determinePackage(x)
}
\arguments{
\item{x}{quoted expression}
}
\description{
Determine which graph package evaluation will
occur in.
}
\details{
Fails if functions match distinct
packages.  Might need to change this in future if
different packages work with the same format.
}
