% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Family.R
\name{districts}
\alias{districts}
\alias{neighbourhoods,}
\alias{un}
\alias{neighbourhoods}
\alias{cliques}
\title{Familial Mixed Graph Groups}
\usage{
districts(graph)

un(graph, sort = 1)

neighbourhoods(graph)

cliques(graph)
}
\arguments{
\item{graph}{\code{mixedgraph} object}

\item{sort}{should vertices be sorted?}
}
\description{
The usual familial relations between vertices in
mixed graphs.
}
\details{
\code{districts} and \code{neighbourhoods} find the
bidirected-connected and undirected-connected components of \code{graph}.
\code{un} finds the undirected part of \code{graph}.

\code{cliques} uses the Bron-Kirbosch algorithm to find 
maximal connected subsets.
}
\section{Functions}{
\itemize{
\item \code{un}: Obtain undirected component

\item \code{neighbourhoods}: Obtain neighbourhoods

\item \code{cliques}: Obtain maximal complete undirected subsets
}}

