% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual.R
\name{dual}
\alias{dual}
\title{Dual Graphs}
\usage{
dual(graph)
}
\arguments{
\item{graph}{a mixed graph object}
}
\description{
Constructs the (undirected) dual graph for a mixed graph
}
\details{
Uses the skeleton and then inverts the set of edges
}
