% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphs.R
\name{edgeTypes}
\alias{edgeTypes}
\alias{vertexTypes}
\title{See list of vertex and edge types}
\usage{
edgeTypes()

vertexTypes()
}
\description{
Returns data frame of current edge types.
}
\section{Functions}{
\itemize{
\item \code{vertexTypes}: See vertex types
}}

