% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Edges.R
\name{eList}
\alias{eList}
\title{Edge list}
\usage{
eList(edges, directed = FALSE)
}
\arguments{
\item{edges}{\code{adjList}, \code{edgeMatrix} or \code{adjMatrix}}

\item{directed}{logical: if TRUE edges are assumed directed}
}
\description{
Returns an \code{eList} object from an adjacency matrix
or edge matrix object.
}
\details{

}
