% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solver.R
\name{glmLasso}
\alias{glmLasso}
\title{glmLasso is used to fit models with lasso penalty.}
\usage{
glmLasso(X, y, o = NULL, lambda = 1, family = "gaussian", KLB = NULL,
  thresh = NULL, max.iter = 1e+08, threads = NULL)
}
\arguments{
\item{X}{is a n*p input matrix.}

\item{y}{is the response vector (n elements).}

\item{o}{is the offset vector (n elements).}

\item{lambda}{is the penalty weight vector, can also be a single value, we may apply the scalar to a vector in our function.}

\item{family}{is a description of the error distribution and link function to be used in the model. In our package, "binomial", "gaussian" and  "poisson" are available.}

\item{KLB}{is the iterations number which the support set unchanged.}

\item{thresh}{is the precision when the solver to stop optimizing, when the KLB and thresh are setted at the same time, the solver would stop when any of them is achieved.}

\item{max.iter}{is the maximum number of iterations to be performed for the optimization.}

\item{threads}{is the number of threads used for paralleling the ADMM process, the default value would be the available cores on a machine.}
}
\value{
the coefficients vector
}
\description{
glmLasso is used to fit models with lasso penalty.
}
\examples{
X <- matrix(rnorm(500), ncol = 10)
y <- rbinom(50, 1, 0.6)
o <- rnorm(50)
lambda <- c(1:10)/10
glmLasso(X, y, o, lambda, family = "binomial", KLB = 10, thresh = 0.5, max.iter = 1e7, threads = 4)

}
