% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solver.R
\name{glmRidge}
\alias{glmRidge}
\title{glmRidge is used to fit models with ridge penalty.}
\usage{
glmRidge(X, y, o = NULL, lambda = 1, family = "gaussian",
  thresh = 0.001, max.iter = 1e+06)
}
\arguments{
\item{X}{is a n*p input matrix.}

\item{y}{is the response vector (n elements).}

\item{o}{is the offset vector (n elements).}

\item{lambda}{is a single value of ridge penalty.}

\item{family}{is a description of the error distribution and link function to be used in the model. In our package, "binomial", "gaussian" and  "poisson" are available.}

\item{thresh}{indicates when to stop the solver. glmRidge will check the Euclidean norm of the change.}

\item{max.iter}{is the maximum number of iterations to be performed for the optimization.}
}
\value{
the coefficients vector
}
\description{
glmRidge is used to fit models with ridge penalty.
}
\examples{
X <- matrix(rnorm(500), ncol = 10)
y <- rbinom(50, 1, 0.6)
o <- rnorm(50)

glmRidge(X, y, o, lambda = 0.5, family = "logistic", thresh = 0.005, max.iter = 1e5)

}
