% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gr1}
\alias{gr1}
\alias{gr2}
\alias{grIV}
\alias{grVerma}
\title{Example Graphs}
\format{These are all \code{mixedgraph} objects.}
\usage{
gr1

gr2

grIV

grVerma
}
\description{
Some example graphs for illustration.
}
\section{Functions}{
\itemize{
\item \code{gr2}: another graph

\item \code{grIV}: The instrumental variables graph

\item \code{grVerma}: The Verma graph
}}

\keyword{datasets}
