% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convert.R
\name{\%G\%}
\alias{\%G\%}
\title{Automatically convert and apply graph function}
\usage{
graph \%G\% .f
}
\arguments{
\item{graph}{a graph object that can be handled
by \code{MixedGraphs}}

\item{.f}{right-hand side of a pipe to be evaluated

This is a version of the pipe `%>%` from \code{magrittr}
that converts the left-hand side into a graph of suitable
format to be used by the right-hand side.}
}
\description{
Experimental automatic conversion function
}
\examples{
gr1 <- graphCr("1 -> 2 <- 3 -> 4")
# use the degree function from the graph package
gr1 \%G\% graph::degree
# equivalently in this case:
gr1 \%>\% 
    convert(format="graphNEL") \%>\%
    degree
    
}
