% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_algs.R
\name{perfect_elim_order}
\alias{perfect_elim_order}
\alias{peo}
\alias{n_fill_in}
\title{Algorithms for common graphical combinatorial problems}
\usage{
perfect_elim_order(graph, v)

peo(graph, v)

n_fill_in(graph, v, any = FALSE)
}
\arguments{
\item{graph}{object of class \code{mixedgraph}}

\item{v}{subset of vertices}

\item{any}{logical, if \code{TRUE} returns \code{TRUE} if any edge is added, \code{FALSE} otherwise}
}
\description{
Algorithms for common graphical combinatorial problems
}
\details{
Defaults to the entire graph if \code{v} is not supplied.  This algorithm applies
only to undirected edges.
}
\section{Functions}{
\itemize{
\item \code{perfect_elim_order()}: find a perfect elimination ordering (quadratic)

\item \code{peo()}: find a perfect elimination ordering

\item \code{n_fill_in()}: number of edges to add for given ordering

}}
