% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Edges.R, R/Graphs.R
\name{nedge}
\alias{nedge}
\alias{graphOps}
\alias{vnames}
\alias{nv}
\title{Graph Operations}
\usage{
nedge(graph, edges)

vnames(graph)

nv(graph)
}
\arguments{
\item{graph}{an object of class \code{mixedgraph}}

\item{edges}{character vector of edge types to include, defaults to all}
}
\description{
Graph Operations
}
\details{
\code{nedge} uses an internal function \code{nedge2} to count for
each type of edge separately. Do we want this to apply to graphs or edge lists
or both?
}
\section{Functions}{
\itemize{
\item \code{nedge}: number of edges

\item \code{vnames}: names of vertices for \code{mixedgraph} object

\item \code{nv}: number of vertices for \code{mixedgraph} object
}}

