% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADMGs.R
\name{headsTails}
\alias{headsTails}
\title{Get list of heads and tails}
\usage{
headsTails(graph, r = TRUE, by_district = FALSE, set_list, max_head)
}
\arguments{
\item{graph}{a \code{mixedgraph} object}

\item{r}{logical, should recursive heads be used?}

\item{by_district}{logical, should these be grouped by district?}

\item{set_list}{list of intrinsic sets}

\item{max_head}{largest head size to consider (\code{r=FALSE} only)}
}
\description{
Get list of heads and tails
}
\details{
Returns a list of heads and their corresponding tails.
}
