% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADMGs.R
\name{intrinsicSets}
\alias{intrinsicSets}
\alias{intrinsicSets2}
\alias{intrinsicClosure}
\title{List Intrinsic Sets of ADMG}
\usage{
intrinsicSets(graph, r = TRUE, by_district = FALSE, sort = 2, recall = FALSE)

intrinsicSets2(graph, r = TRUE, by_district = FALSE, maxbarren, sort = 1)

intrinsicClosure(graph, set, r = TRUE, sort = 1)
}
\arguments{
\item{graph}{object of class \code{mixedgraph}, must be an ADMG.}

\item{r}{logical, should recursive head definition be used? Defaults to \code{TRUE}}

\item{by_district}{logical, should intrinsic sets be grouped by district? Defaults to \code{FALSE}}

\item{sort}{should output be sorted/unique?}

\item{recall}{logical: is this a recalling of the function (internal use only)}

\item{maxbarren}{Maximum number of barren nodes (i.e. head size) to consider}

\item{set}{set to find intrinsic closure of}
}
\description{
List Intrinsic Sets of an ADMG, possibly by district
}
\details{
\code{intrinsicSets} returns a list of integer vectors, 
each being an intrinsic set (or if \code{by_district = TRUE} 
a list of lists, each containing the intrinsic sets in a single 
district). If \code{r = FALSE} the intrinsic sets are the heads 
and their dis-tails.

\code{intrinsicClosure} returns an integer vector containing the closure of set.
}
\section{Functions}{
\itemize{
\item \code{intrinsicSets2}: Alternative method for non-recursive heads only

\item \code{intrinsicClosure}: Get the intrinsic closure of a set
}}

