% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphs.R
\name{is.adjMatrix}
\alias{is.adjMatrix}
\title{Check if object could be an adjacency matrix}
\usage{
is.adjMatrix(object, n)
}
\arguments{
\item{object}{purported adjMatrix}

\item{n}{number of vertices in graph}
}
\description{
Currently assumes that entries must be non-negative
}
