% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Types.R
\name{is.cyclic-deprecated}
\alias{is.cyclic-deprecated}
\alias{is.cyclic}
\title{Test cyclicity of graph}
\usage{
is.cyclic(A, B, C = integer(0), n = max(c(A, B, C)))
}
\arguments{
\item{graph}{\code{mixedgraph} object}
}
\description{
Looks for cycles among edges of type 'directed'.  Not tested.
Maybe should add option to not test among all directed edges types.
}
