% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Types.R
\name{is.cyclic}
\alias{is.cyclic}
\title{Test cyclicity of graph}
\usage{
is.cyclic(graph)
}
\description{
Looks for cycles among edges of type 'directed'.  Not tested.
Maybe should add option to not test among all directed edges types.
}

