% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Edges.R
\name{is.edgeMatrix}
\alias{is.edgeMatrix}
\alias{is.eList}
\alias{is.adjMatrix}
\alias{is.adjList}
\title{Check if object could be an edgeMatrix}
\usage{
is.edgeMatrix(object, n, checknm = TRUE)

is.eList(object, n, checknm = TRUE)

is.adjMatrix(object, n, checknm = TRUE)

is.adjList(object, n, checknm = TRUE)
}
\arguments{
\item{object}{purported edgeMatrix}

\item{n}{(optionally) number of vertices in graph}

\item{checknm}{logical: use class of object to determine answer?}
}
\description{
Check if object could be an edgeMatrix
}
\section{Functions}{
\itemize{
\item \code{is.eList()}: Check if object could be eList

\item \code{is.adjMatrix()}: Check if object could be adjMatrix

\item \code{is.adjList()}: Check if object could be adjList

}}
