% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphs.R
\name{is.mixedgraph}
\alias{is.mixedgraph}
\title{Test for \code{mixedgraph}}
\usage{
is.mixedgraph(object)
}
\arguments{
\item{object}{Object to be tested}
}
\description{
Is this object a \code{mixedgraph}?
}
\details{
Returns \code{TRUE} or \code{FALSE}
}
