% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adjacencies.R
\name{is_complete}
\alias{is_complete}
\title{Check if a subgraph is complete}
\usage{
is_complete(graph, v)
}
\arguments{
\item{graph}{an object of class \code{mixedgraph}}

\item{v}{set of vertices to test completeness of}
}
\description{
Check if a subgraph is complete
}
