% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_algs.R
\name{list_isomorphisms}
\alias{list_isomorphisms}
\title{Obtain isomorphisms from one graph to another}
\usage{
list_isomorphisms(graph1, graph2, use_skel = TRUE)
}
\arguments{
\item{graph1, graph2}{two graphs to consider}

\item{use_skel}{logical indicating whether to take the skeleton}
}
\description{
List mappings that map one undirected graph onto another
}
