% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_sep.R
\name{m_sep}
\alias{m_sep}
\title{Test m-separation}
\usage{
m_sep(graph, A, B, C, check = TRUE)
}
\arguments{
\item{graph}{an object of class \code{mixedgraph}}

\item{A, B, C}{sets of vertices in \code{graph}}

\item{check}{logical: check if \code{graph} is a summary graph}
}
\description{
Check if A is m-separated from B by C in a summary graph \code{graph}.
}
