% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Edges.R
\name{makeEdgeList}
\alias{makeEdgeList}
\title{Make an edgeList from a collection of edges}
\usage{
makeEdgeList(...)
}
\arguments{
\item{...}{list of named edge objects}
}
\description{
Make an edgeList from a collection of edges
}
