% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutilate.R
\name{match_vnames}
\alias{match_vnames}
\title{Function to match variables using their names}
\usage{
match_vnames(graph, edges)
}
\arguments{
\item{graph}{an object of class \code{mixedgraph}}

\item{edges}{an \code{edgeList} object created by \code{edgeCr}}
}
\description{
Function to match variables using their names
}
