% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Family.R
\name{mb}
\alias{mb}
\title{Find Markov blanket}
\usage{
mb(graph, v, A, check = TRUE, sort = 1)
}
\arguments{
\item{graph}{\code{mixedgraph} object}

\item{v}{a vertex, childless in \code{A}}

\item{A}{an ancestral collection of vertices}

\item{check}{logical: check \code{A} is ancestral?}

\item{sort}{integer:1 for unique but unsorted, 2 for
sorted (0 for possibly repeated and unsorted).  If edges are stored as a matrix
then output will always be unique and sorted.}
}
\description{
Find the Markov blanket for a vertex in an
ancestral set
}
\details{
Finds the Markov blanket of \code{v} in \code{A}.
}
