% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Graphs.R
\name{mixedgraph}
\alias{mixedgraph}
\title{Construct a mixed graph}
\usage{
mixedgraph(n, v = seq_len(n), edges = list(), vnames, vtype)
}
\arguments{
\item{n}{integer number of vertices in graph}

\item{v}{integer identifiers of vertices (length \code{n})}

\item{edges}{named list of edges in graph}

\item{vnames}{character vector of vertex names (defaults to \code{x1}, \code{x2}, ...)}

\item{vtype}{optionally, a character vector of vertex types}
}
\description{
Construct a mixed graph
}
\seealso{
\code{\link{graphCr}}.
}

