% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_sep.R
\name{moralize}
\alias{moralize}
\title{Moralize a mixed graph}
\usage{
moralize(graph, C, check = TRUE)
}
\arguments{
\item{graph}{an object of class \code{mixedgraph}}

\item{C}{optional set of vertices to condition upon}

\item{check}{logical: should we check graph is a summary graph?}
}
\description{
Moralizes a mixed graph, possibly conditional on a subset of vertices.
}
\details{
This works for any summary graph, though currently needs the
graph to be the original one (and not any subgraph) in order to be
guaranteed to run correctly.
}
