% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutilate.R
\name{morphEdges}
\alias{morphEdges}
\title{Transform edges to different type}
\usage{
morphEdges(graph, from, to, topOrd)
}
\arguments{
\item{graph}{\code{mixedgraph} object}

\item{from}{character vector of edges to transform (default is all)}

\item{to}{character string of new edge type}

\item{topOrd}{optional topological order for directing edges}
}
\description{
Transform edges to different type
}
\details{
\code{to} must be a single entry
}
