% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutilate.R
\name{mutilate}
\alias{mutilate}
\title{Delete edges}
\usage{
mutilate(graph, A, B, etype, dir = 0L, internal = FALSE)
}
\arguments{
\item{graph}{a \code{mixedgraph} object}

\item{A, B}{sets of vertices in \code{graph}}

\item{etype}{which edges to remove}

\item{dir}{indicates whether only edges of certain orientation are removed}

\item{internal}{logical: should only edges within \code{A} be removed?}
}
\description{
Remove edges adjacent to set of vertices
}
\details{
If no edge type is specified, then all edges are removed.
If \code{dir=1}, then directed edges out of \code{A} are removed,
but ones into \code{A} are preserved; for \code{dir=-1} the reverse,
and for \code{dir=0} (the default), direction is irrelevant.
If a second set \code{B} is specified, then all edges between \code{A}
and \code{B} are removed.

If \code{A} is not specified, then this will remove the edge classes specified in
\code{etype} entirely.

Note that specifying \code{internal=TRUE} and providing a set \code{B} will
result in an error.

Note that specifying \code{internal=TRUE} and providing a set \code{B} will
result in an error.
}
