% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Edges.R
\name{nedge}
\alias{nedge}
\title{Give number of edges}
\usage{
nedge(graph, edges)
}
\arguments{
\item{graph}{an object of class \code{mixedgraph}}

\item{edges}{character vector of edge types to include, defaults to all

Uses an internal function \code{nedge2} to count for
each type of edge separately.  
Do we want this to apply to graphs or edge lists
or both?}
}
\description{
Give number of edges
}
