% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Adjacencies.R
\name{pa}
\alias{anc}
\alias{ant}
\alias{ch}
\alias{dec}
\alias{dis}
\alias{nb}
\alias{pa}
\alias{sp}
\title{Familial Mixed Graph Relations}
\usage{
pa(graph, v, sort = 1)

ch(graph, v, sort = 1)

sp(graph, v, sort = 1)

nb(graph, v, sort = 1)

anc(graph, v, sort = 1)

dec(graph, v, sort = 1)

ant(graph, v, sort = 1)

dis(graph, v, sort = 1)
}
\arguments{
\item{graph}{\code{mixedgraph} object}

\item{v}{collection of vertices}

\item{sort}{integer:1 for unique but unsorted, 2 for
sorted (0 for possibly repeated and unsorted).  If edges are stored as a matrix
then output will always be unique and sorted.}
}
\description{
The usual familial relations between vertices in
mixed graphs.
}
\details{
\code{pa}, \code{ch}, \code{sp} and \code{nb} find the
parents, children, spouses and neighbours of \code{v} respectively.
\code{anc}, \code{dec}, \code{ant}, \code{dis} finds the ancestors
descendants, anterior and district of \code{v} respectively.
}
\section{Functions}{
\itemize{
\item \code{ch}: 

\item \code{sp}: 

\item \code{nb}: 

\item \code{anc}: 

\item \code{dec}: 

\item \code{ant}: 

\item \code{dis}: 
}}

