% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_algs.R
\name{perfect_elim_order}
\alias{perfect_elim_order}
\title{Obtain a perfect elimination order for an undirected graph}
\usage{
perfect_elim_order(graph, v)
}
\arguments{
\item{graph}{object of class \code{mixedgraph}}

\item{v}{subset of vertices whose undirected edges should be oriented}
}
\description{
Obtain a perfect elimination order for an undirected graph
}
\details{
Defaults to the entire graph if \code{v} is not supplied.
}
