% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphs.R
\name{print.mixedgraph}
\alias{print.mixedgraph}
\title{Print a mixedgraph object}
\usage{
\method{print}{mixedgraph}(x, ...)
}
\arguments{
\item{x}{an object of class \code{mixedgraph}}

\item{...}{other arguments to \code{print}}
}
\description{
Prints a \code{mixedgraph} object to the standard output.
}
