% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphs.R
\name{standardizeEdges}
\alias{standardizeEdges}
\title{Write graphs in standard format}
\usage{
standardizeEdges(graph)
}
\arguments{
\item{graph}{\code{mixedgraph} object

Designed to make comparison of graphs easier}
}
\description{
Write graphs in standard format
}
