% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphs.R
\name{standardizeVertices}
\alias{standardizeVertices}
\title{Force graphs to have vertices 1,\dots,k}
\usage{
standardizeVertices(graph)
}
\arguments{
\item{graph}{\code{mixedgraph} object

Designed to make comparison of graphs easier}
}
\description{
Force graphs to have vertices 1,\dots,k
}
