% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Graphs.R
\name{[.mixedgraph}
\alias{[.mixedgraph}
\alias{subGraph}
\title{Take induced vertex subgraph of mixedgraph}
\usage{
\method{[}{mixedgraph}(graph, v, ...)

subGraph(graph, v, drop = FALSE)
}
\arguments{
\item{graph}{a \code{mixedgraph} object}

\item{v}{vertices to keep}

\item{drop}{force drop removed vertices from representation?}
}
\description{
Take induced vertex subgraph of mixedgraph
}
\section{Methods (by generic)}{
\itemize{
\item \code{[}: 
}}

