% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test}
\alias{test}
\title{L1-penalized Poisson Regression
You can see glmPR.default and glmPR.formula for more details}
\usage{
test(X, y, method = "Gaussian", lambda = 0.5)
}
\arguments{
\item{X}{the input matrix}

\item{y}{the responce vector}

\item{method}{the solver method}

\item{lambda}{a constant scalar parameter to control the influence of L1-Norm}

\item{...}{other parameters}
}
\description{
L1-penalized Poisson Regression
You can see glmPR.default and glmPR.formula for more details
}
\examples{
X <- matrix(rnorm(500), ncol = 10)
y <- rpois(50, 3)
glm(y ~ X, family = "poisson")
test(cbind(1, X), y, "poisson", 0.0)

}
