% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Types.R
\name{test_graph-deprecated}
\alias{test_graph-deprecated}
\alias{is.UG}
\alias{is.DAG}
\alias{is.ADMG}
\alias{is.SG}
\title{Deprecated functions to test type of graph}
\usage{
is.UG(graph)

is.DAG(graph)

is.ADMG(graph)

is.SG(graph)
}
\arguments{
\item{graph}{\code{mixedgraph} object

\code{is.UG}, \code{is.DAG}, \code{is.ADMG}, \code{is.SG} respectively
test whether a graph is an undirected graph, directed acyclic graph (DAG), 
acyclic directed mixed graph (ADMG) or summary graph.  These have since 
been replaced by \code{is_UG} etc.}
}
\description{
Deprecated functions to test type of graph
}
\section{Functions}{
\itemize{
\item \code{is.UG}: version of \code{is_UG}

\item \code{is.DAG}: version of \code{is_DAG}

\item \code{is.ADMG}: version of \code{is_ADMG}

\item \code{is.SG}: version of \code{is_SG}
}}

