% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{testNewton}
\alias{testNewton}
\title{Test function for Newton Solver}
\usage{
testNewton(X, y, method = "Gaussian", lambda = 0.5)
}
\arguments{
\item{X}{the input matrix}

\item{y}{the responce vector}

\item{method}{the solver method}

\item{lambda}{a constant scalar parameter to control the influence of L2-Norm}

\item{...}{other parameters}
}
\description{
Test function for Newton Solver
}
\examples{
X <- matrix(rnorm(500), ncol = 10)
y <- rpois(50, 3)
glm(y ~ X, family = "poisson")
testNewton(cbind(1, X), y, "poisson", 0.0)
}
