% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Edges.R
\name{withAdjMatrix}
\alias{withAdjMatrix}
\alias{withAdjList}
\alias{withEdgeMatrix}
\alias{withEdgeList}
\title{Change representation of edges}
\usage{
withAdjMatrix(graph, edges, sparse = FALSE, force = FALSE)

withAdjList(graph, edges, force = FALSE)

withEdgeMatrix(graph, edges, force = FALSE)

withEdgeList(graph, edges, force = FALSE)
}
\arguments{
\item{graph}{an object of class \code{mixedgraph}}

\item{edges}{character vector of edge types to change, defaults to all}

\item{sparse}{logical: should sparse matrices be used?}

\item{force}{logical: should edge sets be added when named?}
}
\description{
Change edge representation of (part of) graph to use \code{adjMatrix},
\code{adjList}, \code{eList} or \code{edgeMatrix} formats.
}
\section{Functions}{
\itemize{
\item \code{withAdjList()}: Change to \code{adjList} format

\item \code{withEdgeMatrix()}: Change to \code{edgeMatrix} format

\item \code{withEdgeList()}: Change to \code{eList} format

}}
