% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Edges.R
\name{withEdgeList}
\alias{withEdgeList}
\title{Edge list representation}
\usage{
withEdgeList(graph, edges)
}
\arguments{
\item{graph}{an object of class \code{mixedgraph}}

\item{edges}{character vector of edge types to change, defaults to all}
}
\description{
Change edge representation of graph to use edge lists
}
