% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutilate.R
\name{addEdges}
\alias{addEdges}
\alias{removeEdges}
\title{Add or remove edges}
\usage{
addEdges(graph, edges, ..., remDup = TRUE)

removeEdges(graph, edges, ..., force = FALSE, fast = FALSE)
}
\arguments{
\item{graph}{a \code{mixedgraph} object}

\item{edges}{list of edges to be added/removed}

\item{...}{edges to be added with arguments matching names of edge types}

\item{remDup}{logical: should we check for duplicate edges?}

\item{force}{should we just ignore edges not actually present?}

\item{fast}{fast version for when graph already uses adjacency matrices}
}
\description{
Add or remove edges
}
\details{
The \code{remDup} argument is set by default to
remove duplicate edges. Currently \code{removeEdges()} forces
all edges to be represented by adjacency matrices.

If the \code{edges} argument to \code{addEdges()} is a character vector, then an
empty edge list is added (if no entry is currently present).

The \code{fast} argument for \code{removeEdges} requires that
the edge to be removed is given as a single vector of length 2.
}
\section{Functions}{
\itemize{
\item \code{removeEdges()}: remove edges

}}
