% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adjacencies.R
\name{adj}
\alias{adj}
\title{Find adjacent vertices}
\usage{
adj(graph, v, etype, dir = 0, inclusive = TRUE, sort = 1, force = FALSE)
}
\arguments{
\item{graph}{an object of class \code{mixedgraph}}

\item{v}{vertices to find adjacencies}

\item{etype}{edge types to consider; defaults to all}

\item{dir}{for directed edges, indicates which direction to search in:
1: along direction, -1: against direction, 0: both directions.}

\item{inclusive}{logical indicating whether elements of \code{v} can be 
included in output group.}

\item{sort}{integer:1 for unique but unsorted, 2 for 
sorted (0 for possibly repeated and unsorted).  If edges are stored as a matrix
then output will always be unique and sorted.}

\item{force}{logical - should invalid \code{v} be ignored?}
}
\description{
Generic function for finding adjacent vertices based on any kind of edge.
}
\details{
The argument \code{directed} is recycled for multiple edge types, but 
has no effect for edges without a specified direction.  If any \code{v} is 
not a vertex of \code{graph}, an error is returned, unless \code{force=TRUE}.
}
\seealso{
\code{\link{grp}} for paths
}
