% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adjacencies.R
\name{adj}
\alias{adj}
\alias{adjacent}
\title{Find adjacent vertices}
\usage{
adj(graph, v, etype, dir = 0, inclusive = TRUE, sort = 1, force = FALSE)

adjacent(graph, v1, v2, etype, dir = 0)
}
\arguments{
\item{graph}{an object of class \code{mixedgraph}}

\item{v}{vertices to find adjacencies}

\item{etype}{edge types to consider; defaults to all}

\item{dir}{for directed edges, indicates which direction to search in:
1: along direction, -1: against direction, 0: both directions.}

\item{inclusive}{logical indicating whether elements of \code{v} can be
included in output group.}

\item{sort}{integer: 1 for unique but unsorted, 2 for
sorted (0 for possibly repeated and unsorted).  If edges are stored as a matrix
then output will always be unique and sorted.}

\item{force}{logical - should invalid \code{v} be ignored?}

\item{v1, v2}{vertices between which to test for adjacency}
}
\description{
Generic function for finding adjacent vertices based on any kind of edge.
}
\details{
The argument \code{directed} is recycled for multiple edge types, but
has no effect for edges without a specified direction.  If any \code{v} is
not a vertex of \code{graph}, an error is returned, unless \code{force=TRUE}.

The function \code{adjacent} tests for a particular adjacency between
two vertices.
}
\section{Functions}{
\itemize{
\item \code{adjacent()}: Test for adjacency

}}
\examples{
adj(gr1, v=1, etype="directed")
adj(gr1, v=1, etype="directed", dir=-1)
adjacent(gr1, 1, 3, etype="directed", dir=1)

}
\seealso{
\code{\link{grp}} for vertices joined by paths, \code{\link{groups}}
for partitions of vertices by adjacency type.
}
