% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Edges.R
\name{adjMatrix}
\alias{adjMatrix}
\title{Get adjacency matrix from list of edges}
\usage{
adjMatrix(edges, n, directed = FALSE)
}
\arguments{
\item{edges}{list or data frame giving edges}

\item{n}{total number of vertices, defaults to maximal value}

\item{directed}{logical: if TRUE edges are assumed directed}
}
\description{
Get adjacency matrix from list of edges
}
