% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adjacencies.R
\name{anSets}
\alias{anSets}
\title{Find ancestral sets of a graph.}
\usage{
anSets(graph, topOrder)
}
\arguments{
\item{graph}{object of class \code{mixedgraph}, should be a summary graph}

\item{topOrder}{optional topological order of vertices}
}
\description{
Find ancestral sets of a graph.
}
\details{
Algorithm:
1. Find a topological order of nodes.
2. Base case: {} is ancestral
3. Induction: (i) Assume we have a list L of all ancestral sets involving Xi-1 in the order.
(ii) If an ancestral set S in L contains all parents of Xi, Xi + S is also ancestral.
}
\author{
Ilya Shpitser
}
