% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Family.R
\name{barren}
\alias{barren}
\alias{orphaned}
\alias{sterile}
\title{Find barren, sterile, orphaned vertices}
\usage{
barren(graph, v = graph$v)

orphaned(graph, v = graph$v)

sterile(graph, v = graph$v)
}
\arguments{
\item{graph}{an object of class \code{mixedgraph}}

\item{v}{set of vertices of \code{graph}}
}
\description{
Find barren, sterile, orphaned vertices
}
\details{
Barren vertices (within \code{v}) are those
that have no proper descendants also within \code{v}.
Sterile (orphaned) vertices in \code{v} have no
children (parents) also within \code{v}.
}
\section{Functions}{
\itemize{
\item \code{orphaned()}: find vertices with no parents

\item \code{sterile()}: find vertices with no children in the same set

}}
