% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adjacencies.R
\name{collapse}
\alias{collapse}
\title{Collapse multiple edge types to a single list or matrix}
\usage{
collapse(edges, v1, v2, dir = 1, matrix = FALSE)
}
\arguments{
\item{edges}{list of edges to include}

\item{v1, v2}{incoming and outgoing vertices, defaults to all.}

\item{dir}{direction}

\item{matrix}{logical indicating whether to force the return of an adjacency matrix}
}
\description{
Collapse multiple edge types to a single list or matrix
}
\details{
returns an edgeMatrix or adjacency matrix for possibly multiple edge types.
If any of the edges are specified as an adjacency matrix, then the output will also
be an adjacency matrix.
}
