% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Convert.R
\name{convert}
\alias{convert}
\title{Convert graph to format associated with specific package}
\usage{
convert(graph, format = "mixedgraph", cur_format, ...)
}
\arguments{
\item{graph}{graphical object, of one of classes listed below}

\item{format}{character string giving new format to convert to}

\item{cur_format}{character string of current format; can be deduced
from the object in many cases}

\item{...}{}
}
\description{
Convert graph to format associated with specific package
}
\details{
Currently limited functionality.  Possible formats
are "ADMG", "ggm", "graphNEL", "graphAM", "PAG", "bn", "igraph",
and of course "mixedgraph".
The function can recognise the class of objects other than
"ggm", which is just an adjacency matrix.

Implemented thus far:
\itemize{
\item mixedgraph <-> ggm
\item mixedgraph <-> ADMG
\item mixedgraph <-> graphNEL
\item mixedgraph <-> graphAM
\item mixedgraph <-- PAG
\item mixedgraph <-> bn
\item mixedgraph <-- igraph
}

\code{ggm} entries must be specified by \code{cur_format = "ggm"}.
}

