% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convert.R
\name{convert}
\alias{convert}
\title{Convert graph to format associated with specific package}
\usage{
convert(graph, format = "mixedgraph", cur_format, ...)
}
\arguments{
\item{graph}{graphical object, of one of classes listed below}

\item{format}{character string giving new format to convert to}

\item{cur_format}{character string of current format; can be deduced 
from the object in many cases}

\item{...}{}
}
\description{
Convert graph to format associated with specific package
}
\details{
Possible formats
are 
\itemize{
\item \code{\link{mixedgraph}}
\item \code{ggm} an adjacency matrix as specified in the \code{ggm} package; 
\item \code{graphNEL}, \code{graphAM}, \code{graphBAM} from the \code{graph} package;
\item \code{igraph};
\item \code{PAG}: that is, the output of \code{pc()} or \code{fci()} functions in the \code{pcalg} package;
\item \code{bn} from the \code{bnlearn} package;
\item \code{ADMG}, from the \code{ADMGs} package.
}
The function can recognise the class of objects other than
"ggm", which is just an adjacency matrix.

Implemented natively thus far:
\itemize{
\item mixedgraph <-> ggm
\item mixedgraph <-> ADMG
\item mixedgraph <-> graphNEL
\item mixedgraph <-> graphAM
\item mixedgraph <-> graphBAM
\item mixedgraph <-> PAG
\item mixedgraph <-> bn
\item mixedgraph <-> igraph (goes mixedgraph -> graphNEL -> igraph)
\item graphNEL <-> igraph (using functions in the \code{igraph} package)
\item graphNEL, graphAM <-> bn (using functions in the \code{bnlearn} package)
}

\code{ggm} entries must be specified by \code{cur_format = "ggm"}.  
\code{PAG} objects are those whose class inherits from \code{gAlgo}, 
or an adjacency matrix if \code{cur_format = "PAG"} is specified.
}

