% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Edges.R
\name{edgeList}
\alias{edgeList}
\title{Deprecated function}
\usage{
edgeList(edges, directed = FALSE)
}
\arguments{
\item{edges}{an "eList" object}

\item{directed}{logical, should edges be interpreted as directed?}
}
\description{
Deprecated function
}
