% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Edges.R
\name{edgeList}
\alias{edgeList}
\title{Edge list}
\usage{
edgeList(edges, directed = FALSE)
}
\arguments{
\item{edges}{}

\item{directed}{logical: if TRUE edges are assumed directed}
}
\description{
Edge list
}

