% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Edges.R
\name{edgeMatrix}
\alias{edgeMatrix}
\title{Get edges from adjacency matrix or list}
\usage{
edgeMatrix(edges, directed = FALSE)
}
\arguments{
\item{directed}{logical: if TRUE edges are assumed directed}
}
\description{
Get edges from adjacency matrix or list
}

