% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gr1}
\alias{gr1}
\alias{gr2}
\alias{grIV}
\alias{grVerma}
\alias{grJamie}
\title{Example Graphs}
\format{
These are all \code{mixedgraph} objects.

An object of class \code{mixedgraph} of length 3.

An object of class \code{mixedgraph} of length 3.

An object of class \code{mixedgraph} of length 3.

An object of class \code{mixedgraph} of length 3.
}
\usage{
gr1

gr2

grIV

grVerma

grJamie
}
\description{
Some example graphs for illustration.
}
\section{Functions}{
\itemize{
\item \code{gr2}: another graph

\item \code{grIV}: The instrumental variables graph

\item \code{grVerma}: The Verma graph

\item \code{grJamie}: Jamie's 'double' Verma graph
}}

\keyword{datasets}
