% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphs.R
\name{graphCr}
\alias{graphCr}
\title{Create graph from character parsing}
\usage{
graphCr(char, ..., useMatrices = FALSE, representation = "mixedgraph")
}
\arguments{
\item{char}{string of inputs given by vertex names separated by edges}

\item{...}{other strings of further edges}

\item{useMatrices}{in mixed graph representation, should the output
use adjacency matrices?}

\item{representation}{}
}
\description{
Convenience function for quickly constructing graphs with character input
}
\details{
Symbols \code{-<>=*|:} are assumed to be part of an edge, so
cannot be used in node names for this function.
}
\examples{
graphCr("1--->2<-->3<-4","2<->4,4->5")
graphCr("1-2-3-4-1", representation="graphNEL")  # requires package 'graph'
}

