% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphs.R
\name{graphCr}
\alias{graphCr}
\title{Create graph from character parsing}
\usage{
graphCr(
  char,
  ...,
  mode = "adjList",
  useMatrices = FALSE,
  format = "mixedgraph"
)
}
\arguments{
\item{char}{string of inputs given by vertex names separated by edges}

\item{...}{other strings of further edges}

\item{mode}{format for edges in graph}

\item{useMatrices}{deprecated argument: in \code{mixedgraph} representation, should the output
use adjacency matrices?}

\item{format}{type of graph format to use, options are \code{mixedgraph} 
(the default), \code{graphNEL} (and \code{graphAM}, \code{graphBAM}), 
\code{igraph}, \code{ggm}, \code{bn}, \code{PAG}.}
}
\description{
Convenience function for quickly constructing graphs with character input
}
\details{
Symbols \code{-<>=*|:} are assumed to be part of an edge, so
cannot be used in node names using this function.
}
\examples{
graphCr("1--->2<-->3<-4","2<->4,4->5")
graphCr("1-2-3-4-1", representation="graphNEL")  # requires package 'graph'
}
