% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphs.R
\name{graphCr}
\alias{graphCr}
\alias{edgeCr}
\title{Create graph from character parsing}
\usage{
graphCr(
  char,
  ...,
  mode = "adjList",
  useMatrices = FALSE,
  format = "mixedgraph"
)

edgeCr(char, mode = "eList")
}
\arguments{
\item{char}{string of inputs given by vertex names separated by edges}

\item{...}{other strings of further edges}

\item{mode}{format for edges in graph}

\item{useMatrices}{deprecated argument: in \code{mixedgraph} representation, should the output
use adjacency matrices?}

\item{format}{type of graph format to use, options are \code{mixedgraph}
(the default), \code{graphNEL} (and \code{graphAM}, \code{graphBAM}),
\code{igraph}, \code{ggm}, \code{bn}, \code{PAG}.}
}
\description{
Convenience function for quickly constructing graphs with character input
}
\details{
Symbols \verb{-<>=*|:} are assumed to be part of an edge, so
cannot be used in node names using this function.  Note that if we want
an edge with an \code{o} on the end (e.g. \verb{o->}) then we must
leave a space between the \code{o} and the variable name.

The \code{edgeCr()} creates just an \code{edgeList} object, and is helpful
for adding edges to existing graphs.
}
\section{Functions}{
\itemize{
\item \code{edgeCr()}: Create edge list by hand

}}
\examples{
graphCr("1--->2<-->3<-4","2<->4,4->5")
graphCr("1-2-3-4-1", representation="graphNEL")  # requires package 'graph'
}
