% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Types.R
\name{is.DAG}
\alias{is.DAG}
\title{Test type of graph}
\usage{
is.DAG(graph)
}
\arguments{
\item{graph}{\code{mixedgraph} object

\code{is.DAG}, \code{is.SG}, \code{is.ADMG} 
test whether a graph is a directed acyclic graph, 
summary graph or acyclic directed mixed graph 
respectively.}
}
\description{
Test type of graph
}

