% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Types.R
\name{is.ADMG}
\alias{is.ADMG}
\alias{is.DAG}
\alias{is.SG}
\title{Test type of graph}
\usage{
is.ADMG(graph)

is.SG(graph)

is.DAG(graph)
}
\arguments{
\item{graph}{\code{mixedgraph} object

\code{is.DAG}, \code{is.SG}, \code{is.ADMG}
test whether a graph is a directed acyclic graph,
summary graph or acyclic directed mixed graph
respectively.}
}
\description{
Test type of graph
}
\section{Functions}{
\itemize{
\item \code{is.ADMG}: 

\item \code{is.SG}: 
}}

