% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Graphs.R
\name{makeGraphComplete}
\alias{makeGraphChain}
\alias{makeGraphComplete}
\alias{makeGraphCycle}
\alias{makeGraphStar}
\title{Graph Constructor Functions}
\usage{
makeGraphComplete(n, type = "undirected")

makeGraphChain(n, type = "undirected")

makeGraphCycle(n, type = "undirected")

makeGraphStar(n, type = "undirected")
}
\arguments{
\item{n}{number of vertices}

\item{type}{type of edges to use}
}
\description{
Graph Constructor Functions
}
\details{
Provides some construction
functions for common graphs: complete graphs,
chains, cycles and stars.
}
\section{Functions}{
\itemize{
\item \code{makeGraphChain}: 

\item \code{makeGraphCycle}: 

\item \code{makeGraphStar}: 
}}

