% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Construct.R
\name{makeGraphComplete}
\alias{makeGraphComplete}
\alias{makeGraphEmpty}
\alias{makeGraphChain}
\alias{makeGraphCycle}
\alias{makeGraphStar}
\alias{makeGraphGrid}
\title{Graph Constructor Functions}
\usage{
makeGraphComplete(n, type = "undirected")

makeGraphEmpty(n)

makeGraphChain(n, type = "undirected")

makeGraphCycle(n, type = "undirected")

makeGraphStar(n, type = "undirected", out = FALSE)

makeGraphGrid(n, m = n, type = "undirected")
}
\arguments{
\item{n}{number of vertices (or rows for a grid)}

\item{type}{type of edges to use}

\item{out}{should edges be directed out from the centre?}

\item{m}{number of columns}
}
\description{
Graph Constructor Functions
}
\details{
Provides some construction 
functions for common graphs: complete graphs, 
chains, cycles and stars.
}
\section{Functions}{
\itemize{
\item \code{makeGraphEmpty}: graph with no edges

\item \code{makeGraphChain}: graph with chain of edges

\item \code{makeGraphCycle}: graph with cycle of edges

\item \code{makeGraphStar}: star shaped graph

\item \code{makeGraphGrid}: grid of vertices
}}

