% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphs.R
\name{makeGraphComplete}
\alias{makeGraphComplete}
\title{Graph Constructor Functions}
\usage{
makeGraphComplete(n, type = "undirected")
}
\arguments{
\item{n}{number of vertices}

\item{type}{type of edges to use}
}
\description{
Graph Constructor Functions
}
\details{
Provides some construction 
functions for common graphs: complete graphs, 
chains, cycles and stars.
}

