% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_sep.R
\name{moralize}
\alias{moralize}
\title{Moralize}
\usage{
moralize(graph, C, check = TRUE)
}
\arguments{
\item{graph}{an object of class \code{mixedgraph}}

\item{C}{optional set of vertices to condition upon}

\item{check}{logical: should we check graph is a summary graph?}
}
\description{
Moralize
}
