% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mutilate.R
\name{mutilate}
\alias{mutilate}
\title{Delete edges}
\usage{
mutilate(graph, A, etype, dir = 0L)
}
\arguments{
\item{graph}{a \code{mixedgraph} object}

\item{A}{a set of vertices in \code{graph}}

\item{etype}{which edges to remove}

\item{directed}{indicates whether only edges of certain orientation are removed}
}
\description{
Remove edges adjacent to set of vertices
}
\details{
If no edge type is specified, then all edges are removed.
If \code{directed=1}, then directed edges out of \code{A} are removed,
but ones into \code{A} are preserved; for \code{directed=-1} the reverse,
and for \code{directed=0} (the default), direction is irrelevant.
}

