% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adjacencies.R
\name{pa}
\alias{pa}
\title{Familial Mixed Graph Relations}
\usage{
pa(graph, v, sort = 1)
}
\arguments{
\item{graph}{\code{mixedgraph} object}

\item{v}{collection of vertices}

\item{sort}{integer:1 for unique but unsorted, 2 for 
sorted (0 for possibly repeated and unsorted).  If edges are stored as a matrix
then output will always be unique and sorted.}
}
\description{
The usual familial relations between vertices in
mixed graphs.
}
\details{
\code{pa}, \code{ch}, \code{sp} and \code{nb} find the 
parents, children, spouses and neighbours of \code{v} respectively.
\code{anc}, \code{dec}, \code{ant}, \code{dis} finds the ancestors
descendants, anterior and district of \code{v} respectively.
}

