% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutilate.R
\name{remove_duplicate_edges}
\alias{remove_duplicate_edges}
\title{Remove duplicate edges}
\usage{
remove_duplicate_edges(edgeList, directed = TRUE, sort = FALSE)
}
\arguments{
\item{edgeList}{\code{edgeList} object}

\item{directed}{are edges directed? (i.e.
does the order of vertices matter?)}

\item{sort}{should vertices in undirected edges
be sorted as a side-effect of the function? (list form
only)

Check for and remove duplicate edges in a list of
edges or edgeMatrix.  Note that this will sort 
elements in}
}
\description{
Remove duplicate edges
}

