% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphs.R
\name{subGraph}
\alias{subGraph}
\title{Take induced vertex subgraph of mixedgraph}
\usage{
subGraph(graph, v, drop = FALSE)
}
\arguments{
\item{graph}{a \code{mixedgraph} object}

\item{v}{vertices to keep}

\item{drop}{force drop removed vertices from representation?}
}
\description{
Take induced vertex subgraph of mixedgraph
}

