% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphs.R
\name{[.mixedgraph}
\alias{[.mixedgraph}
\alias{subGraph}
\title{Take induced vertex subgraph of mixedgraph}
\usage{
\method{[}{mixedgraph}(graph, v, w, ..., drop = FALSE, etype, order = FALSE)

subGraph(graph, v, w, drop = FALSE, etype, order = FALSE)
}
\arguments{
\item{graph}{a \code{mixedgraph} object}

\item{v}{vertices to keep}

\item{w}{other set of vertices for bipartite subgraph}

\item{...}{other arguments}

\item{drop}{force removed vertices to be dropped from representation in adjacency matrices and lists?}

\item{etype}{edge types to keep (defaults to all)}

\item{order}{logical: force graph to follow new order implied by \code{v}?}
}
\description{
Take induced vertex subgraph of mixedgraph
}
\details{
If \code{order} is \code{TRUE}, then \code{edgeMatrix} and \code{eList} 
formats are converted to \code{edgeList}s (\code{adjMatrix} format is preserved).

Note that \code{order} will have no effect if \code{w} is specified.
}
\section{Functions}{
\itemize{
\item \code{[}: bracket notation for subgraphs
}}

