% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphs.R
\name{[.mixedgraph}
\alias{[.mixedgraph}
\alias{subGraph}
\title{Take induced vertex subgraph of mixedgraph}
\usage{
\method{[}{mixedgraph}(graph, v, ..., drop = FALSE, etype)

subGraph(graph, v, drop = FALSE, etype)
}
\arguments{
\item{graph}{a \code{mixedgraph} object}

\item{v}{vertices to keep}

\item{...}{other arguments}

\item{drop}{force removed vertices to be dropped from representation in adjacency matrices?}

\item{etype}{edge types to keep (defaults to all)}
}
\description{
Take induced vertex subgraph of mixedgraph
}
\section{Methods (by generic)}{
\itemize{
\item \code{[}: bracket notation for subgraphs
}}

