% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Types.R
\name{test_graph}
\alias{test_graph}
\alias{is_UG}
\alias{is_DAG}
\alias{is_ADMG}
\alias{is_SG}
\alias{is.UG}
\alias{is.DAG}
\alias{is.ADMG}
\alias{is.SG}
\title{Test type of graph}
\usage{
is_UG(graph)

is_DAG(graph)

is_ADMG(graph)

is_SG(graph)

is.UG(graph)

is.DAG(graph)

is.ADMG(graph)

is.SG(graph)
}
\arguments{
\item{graph}{\code{mixedgraph} object

\code{is_UG}, \code{is_DAG}, \code{is_ADMG}, \code{is_SG} respectively
test whether a graph is an undirected graph, directed acyclic graph (DAG), 
acyclic directed mixed graph (ADMG) or summary graph.}
}
\description{
Test type of graph
}
\section{Functions}{
\itemize{
\item \code{is_UG}: test if an undirected graph

\item \code{is_DAG}: test if a directed acyclic graph (DAG)

\item \code{is_ADMG}: test if an acyclic directed mixed graph (ADMG)

\item \code{is_SG}: test if a summary graph

\item \code{is.UG}: deprecated version of \code{is_UG}

\item \code{is.DAG}: deprecated version of \code{is_DAG}

\item \code{is.ADMG}: deprecated version of \code{is_ADMG}

\item \code{is.SG}: deprecated version of \code{is_SG}
}}

