% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Types.R
\name{topologicalOrder}
\alias{topologicalOrder}
\alias{isTopological}
\title{Find topological ordering of vertices}
\usage{
topologicalOrder(graph, warn = TRUE)

isTopological(graph, v)
}
\arguments{
\item{graph}{a \code{mixedgraph} object}

\item{warn}{logical: should a cyclic graph return a warning?}

\item{v}{an ordering of the vertices}
}
\description{
Looks for topological ordering of vertices with respect to directed edges.
}
\details{
\code{isTopological} tests for an ordering being
topological.  If the graph is cyclic, it returns \code{NA}, with or
without a warning depending on the value of \code{warn}.
}
\section{Functions}{
\itemize{
\item \code{isTopological()}: Check if ordering is topological

}}
