% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Types.R
\name{topologicalOrder}
\alias{topologicalOrder}
\title{Find topological ordering of vertices}
\usage{
topologicalOrder(graph)
}
\description{
Looks for topological ordering of vertices with respect to directed edges.
Not tested
}

