% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Edges.R
\name{withAdjMatrix}
\alias{withAdjMatrix}
\title{Adjacency matrix representation}
\usage{
withAdjMatrix(graph, edges)
}
\arguments{
\item{graph}{an object of class \code{mixedgraph}}

\item{edges}{character vector of edge types to change, defaults to all}
}
\description{
Change edge representation of graph to use adjacency matrices
}
